Rewrite VecWriter to not use Base internals
Turns out we can't actually back VecWriter with a Vector without abusing Vector
internals gruesomely. That's a real shame, because I wish it was possible to
get that kind of control over Vector.
Specifically, we would need to be able to write the `.ref` and `.size` fields
of Vector.

Instead, we use a stop-gap solution. We provide a `ByteVector` type which is a
manual re-implementation of `Vector{UInt8}` that behaves exactly the same, then
we document that `ByteVector` may change in the future to be an alias of
`Vector{UInt8}`.